requirements.txt
pytesseract
Pillow
python-docx
PyMuPDF
libreoffice
